Change how updates are done for extensions
Instead of having a copied repository for each extension, which mostly
just holds the same code as the "official" one, the extension will now be
updated from a second remote, which is the "official" one, and then patches
will be applied if needed. This reduces the overhead of updating extensions
and minimizes the update to just fetching the remote, checking out the new
version branch (wmf/...) and applying patches, if available.

This will only be done for extensions listed in the extensions.list and skins
in the skins.list. Extensions, which are not developed by the WMF, will still
need to be updated manually with git pull e.g.